[RPA-841] Changed endpoint to delete users by id (#321)
* feat(runbotics-orchestrator): ✨ changed delete user endpoint to use id

* refactor(runbotics-orchestrator): :recycle: countUserProcesses without native query + adjustments in deleting unassigned processes

* chore: :pushpin: bump app version to 2.8.0-SNAPSHOT.10

---------

+++ runbotics-orchestrator/src/main/java/com/runbotics/service/UserService.java
@@ -229,9 +229,9 @@ public class UserService {
             .map(AdminUserDTO::new);
     }
 
-    public void deleteUser(String login) {
+    public void deleteUser(Long id) {
         userRepository
-            .findOneByLogin(login)
+            .findOneById(id)
             .ifPresent(
                 user -> {
                     userRepository.delete(user);
@@ -239,7 +239,9 @@ public class UserService {
                 }
             );
 
-        processRepository.deleteUnassignedPrivateProcesses();
+        if (processRepository.countUserProcesses(id) > 0) {
+            processRepository.deleteUnassignedPrivateProcesses();
+        }
     }
 
     public User saveUser(User user) {
